services/document_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
pydantic-settings
python-dotenv
python-multipart
Pillow
docling
vietocr
openai
torch
transformers
celery[redis]
